# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a documentation for the Swipe API",
        "title": "Swipe Public API v0.0.1",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/categories": {
            "get": {
                "description": "get categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "parameters": [
                    {
                        "type": "number",
                        "description": "limit number",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.Slices-entity_Categories"
                        }
                    }
                }
            },
            "post": {
                "description": "insert new category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "parameters": [
                    {
                        "description": "Categories Request",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Categories"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            }
        },
        "/collections": {
            "get": {
                "description": "get collections",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "position of collections",
                        "name": "position",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "limit of cards carousel",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.Article"
                        }
                    }
                }
            },
            "post": {
                "description": "create collections",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "parameters": [
                    {
                        "description": "collections Request",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UploadArticle"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.CollectionUpload"
                        }
                    }
                }
            }
        },
        "/collections/headline": {
            "get": {
                "description": "get list of headline banner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "position of collections",
                        "name": "position",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit headline of collections",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.HeadlineBanners"
                        }
                    }
                }
            },
            "post": {
                "description": "create headline banner into collections",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "parameters": [
                    {
                        "description": "headline banner data request",
                        "name": "banner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.HeadlineBanner"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            }
        },
        "/collections/img": {
            "put": {
                "description": "update collections image",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "parameters": [
                    {
                        "type": "file",
                        "description": "image of collections",
                        "name": "img",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "collections identifier",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            }
        },
...
        "/suppliers": {
            "get": {
                "description": "get suppliers information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit number of suppliers",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.Slices-entity_Suppliers"
                        }
                    }
                }
            },
            "post": {
                "description": "insert new suppliers information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "parameters": [
                    {
                        "description": "Suppliers Request",
                        "name": "Supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.Supplier"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.Article": {
            "type": "object",
            "required": [
                "cards",
                "headline"
            ],
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CardInArticle"
                    }
                },
                "headline": {
                    "type": "string"
                }
            }
        },
        "dtos.CardContent": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "src": {
                    "type": "string"
                }
            }
        },
        "dtos.CardInArticle": {
            "type": "object",
            "required": [
                "category",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CardContent"
                    }
                },
                "src": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dtos.CollectionUpload": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "dtos.HeadlineBanner": {
            "type": "object",
            "required": [
                "body",
                "position"
            ],
            "properties": {
                "body": {
                    "$ref": "#/definitions/dtos.HeadlineBannerBody"
                },
                "created": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        },
        "dtos.HeadlineBannerBody": {
            "type": "object",
            "required": [
                "headline"
            ],
            "properties": {
                "headline": {
                    "type": "string"
                }
            }
        },
        "dtos.HeadlineBannerSlicesBody": {
            "type": "object",
            "required": [
                "headline"
            ],
            "properties": {
                "created": {
                    "type": "string"
                },
                "headline": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "dtos.HeadlineBanners": {
            "type": "object",
            "properties": {
                "headlines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.HeadlineBannerSlicesBody"
                    }
                },
                "position": {
                    "type": "string"
                }
            }
        },
...
        "dtos.OK": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "dtos.Slices-entity_Categories": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Categories"
                    }
                }
            }
        },
        "dtos.Slices-entity_Suppliers": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Suppliers"
                    }
                }
            }
        },
...
        "dtos.Supplier": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "ward": {
                    "type": "string"
                }
            }
        },
...
        "dtos.UploadArticle": {
            "type": "object",
            "required": [
                "cards",
                "headline",
                "position"
            ],
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CardInArticle"
                    }
                },
                "headline": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        },
...
        "entity.Categories": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.Suppliers": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}